Refactored GUI elements out of the "task" file for better re-usability in a future commit.
